Added even more long array missed spots

+++ fNbt/fNbt/Tags/NbtList.cs
@@ -26,7 +26,7 @@ namespace fNbt {
                     if (tags.Count > 0) {
                         throw new ArgumentException("Only empty list tags may have TagType of End.");
                     }
-                }else if (value < NbtTagType.Byte || (value > NbtTagType.IntArray && value != NbtTagType.Unknown)) {
+                }else if (value < NbtTagType.Byte || (value > NbtTagType.LongArray && value != NbtTagType.Unknown)) {
                     throw new ArgumentOutOfRangeException("value");
                 }
                 if (tags.Count > 0) {
